revert to using "hostname -f" by default (partly reverts r4469)
On some systems, e.g. VMs, there may not be a FQDN and Erlang may fail to start with long names without providing it with a full node name in the form of "node@something" (which Erlang always accepts).
Since Java and "hostname -f" determine the same host name, the Java-API library will work out-of-the-box if this method is used on local machines, e.g. with "<node>@localhost" configs.