[branches/0.6] rt_frt: reduce the number of messages sent on lookups
Prior to this change, FRT send a message to a lookup's issuer for each lookup. This change
reduces the amount of messages by only sending a message when the second node in the path
determines that the first hop was suboptimal and that the issuer should know the first
hops successor.