dht_node_move: implement jump operation
Jumping, pratically, is implemented as a leave operation. The successor
who receives the data from the jumper doesn't know about the ongoing
jump. Because of this, during jumps operations, the target id is the
jump target id. During leave operations, the target id is the pred id
or the incremental slide id.