lb_active: let nodes shed more than the average load
Bug in gossip part of the module:

Nodes would shed at most the average global load. This is NOT
the expected behavior as a node should take away as much of its load
as possible. The only exception being that the new nodes' load should
not be lower than the average load.