slide: fix wrongly trying to change the own ID back when aborting a slide pred (regression of r6801)
For any slide, the old own ID is stored and if the slide is aborted, under certain circumstances, we try to restore this ID if it has changed. However if we are aborting a slide with a predecessor, we never changed the own ID. A (simultaneous or interleaving) slide with a successor may have though and changing the ID back is an error.