Makefile: dialyzer plt: tolerate system configurations more flexible
When an Erlang library of our prepared libraray list was not installed
on your system, the building of the dialyzer plt failed. But it is not
mandatory for all libraries to be installed, as it only may raise some
generic warnings during dialyzer runs which are not Scalaris related.

Now, we only include found Erlang libraries in the build plt.