gossip: fix bug in select_data() callback
When returning retry on a select_data() call the trigger message was stored in
the message queue. When contiously returing 'retry' (e.g. in a one node setup with
gossip_cyclon), this led to a self-accelarting endless recursion of resending
and storing the trigger message. Now refactored to ignore the given trigger, i.e.
the current trigger is discarded an the select_data() is called again on the next
again. The return of the select_data() is changed to 'discard_msg' (instead of
'retry') accordingly.

This also fixes the failures in memtest_SUITE.