gossip_SUITE: tweak gossip_load tests
There can be a timing issue preventing correct conversion in the first gossiping
round for test_load. This timing issue is relatively rare. Previously this was
coped with by waiting two rounds per default. This tweak instead repeats the
tests up to three times until success, thus hopefully reducing the standard
execution time of the suite in the default case.

This hopefully also fixes the rare occasions of the test_no_load failing (which
was set to waiting only one round).