routing: remove nodes from external rt
So far, {Node, RTLoopPid} was stored in the external representation of the
routing table. Changed to only store pids. In a best effort approach, the pids
are mostly rt_loop pids, but might also be dht_node pids. Pids of dht_nodes
are only taken if the no rt_loop pid is available. This is the case if the a
node is not present in the internal routing table but added to the external rt
for example from the neighborhood.
Consequently, when sending messages using the external rt, {group_member,
process} has to be used.