Fixed issue with ID generation for user defined ObjectIDClasses
The ID string for objects for which the datastore must generate parts
of the ID (IdGeneratorStrategy.XXX) and which also have an user defined ObjectIdClass
is no longer build by the datastore. This required assumptions about the expected
structure of the ID string which were not true in some cases.