rrepair, merkle: allow splitting the sync range into several Merkle trees
This reduces the amount of unnecessarily sent hashes (and thus message rounds) from the top of the tree when differences occur. On the other hand, communication overhead is slightly increased with very low differences. The selected value of 32 trees increases the costs of transmitting only the first level from ~264 bytes to ~384 bytes (100000 items, P1E=0.01) and is thus acceptable for the gain when differences occur.