comm_connection: fix possible exception in send_internal
Exception (error:function_clause in gen_tcp:send) was seen in a large
scale deployment.  When the tcp connect fails (>= 3 retries), the socket
may not be ready for sending with gen_tcp:send.

There was also a type error: 'fail' (returned by new_connection) was
not allowed for the socket in the state. Now we use the 'notconnected'
more consistently instead of 'fail'.