pid_groups: find_a better shuffles pids
when a client process wanted find a process by name for the first time,
the ets table is queried and the first match was returned (and the
remaining ones were cached and given in a round-robin manner). This
leads to the unintended behaviour that when two new processes query
for the same process name, both get the same pid as first result.
We want to have more randomness in this scenario, so we now shuffle the
list of pids before putting them to the cache.