rbrcseq: Ignore write through infos for round comparisons when collecting replies
If another redundancy scheme than replication is used to store values, the values
returned in the read replies are not required to be the same.
Since the returned value is also included in the write through information of the
last seen round, the equality check with the round of the read reply fails,
causing an endless loop of write throughs.