l_on_cseq: fix recover and merge of passive and active lease
There was a race-condition before this commit. The active lease was recovered with the normal renew process. In parallel the recover triggered a takeover of the passive lease and tried to merge the passive and the active lease. In the error case, the recover of the active lease happened while the merge was going on. This lead to an error in the second merge step with unexpected_owner. The active lease's owner changed to the real owner.