rrepair: deal with duplicates in trivial and shash
Shash was already able to handle duplicates but this was (temporarily) removed
by 2ae4d4b4b6. This commit adds the ability to handle duplicates to both
algorithms by simply sending them to the non-initiator as if the item was
missing/outdated there. If the transfer was wrong, a feedback response will
correct the result.