rrepair: implement an overflow-aware CKV encoding allowing negative deltas
This is e.g. needed to encode duplicates into CKVs since the index of the
second duplicate would be -1. Also, we need duplicates in the CKV so that
merkle_tree can use the info of the number of encoded elements as the number
of available elements on the other node and calculate probabilities
accordingly.