prbr: Optimization for disk backends when writing large values
Optimization for log-structured disk backends when writing
large values. Each time and entry is read (e.g. due to a
normal read operation), its read round is incremented and
the entry is written to the backend. The write-through information
contain the written value. Storing the WTI separate prevents
that the complete value is written repeatedly.