prbr: Stub for collecting round requests
Before a write can be made, a consistent read with a specific round number
must be made. A round requests is the step before the read to retrieve the current
read and write rounds of the replicas. If a quorum of replies was
received the highest known read round number is used for the read.

ToDo: if all replies are consistent, the read request with specific
round number can be skipped