rbrcseq: Prevent sequencing of concurrent read operations
PRBR now stores whether the last round request was part of a
read or write. The read round of an entry does not have to be
incremented for sucessive round requests that are part of reads. Therefore
the read round numbers returned by prbr in such case are consistent
and a qread with explicit round number will not be executed anymore.