prbr: Write is only successful if sequence of writes is gapless
Due to write filter it is possible to define partial write operations.
That requires it to ensure that all replicas see all write operations
without gaps. In the read phase of a write the previous used write round
is delivered by prbr. This must be send back to prbr in the write phase.
If a replica sees a discrepancy the write is denied for that replica.