prbr: Always update proposer UID whenever a round is changed
A round is made unique by its round number and proposer UID.
Previously the proposer UID was only updated in
round_requests. The same proposer UID was used for subsequent
reads or writes. Part of the proposer UID is the request ID
generated by rbrcseq. A read or write which follows a round_
request will have a new request ID.

Whenever a round was changed (incremented), the old proposer UID
of the round_request was used. Certain interleavings of operations
could lead to states in which different replicas have the same
write rounds (number and proposer UID), but diverging values.

(The exact interleaving is unknown, but the problem could be
reproduced consistently for R >= 16 and a high number of
concurrent write processes.)

Always keeping the proposer UID up-to-date whenever a round
is changed for every request sent to prbr fixes this issue.