prbr: Improved handling of concurrent reads
If a round_request belongs to a qread, it is not necessary to increment
any read_rounds. This is because the results of any concurrent qwrite or
qread operations does not change due to a read.
Not incrementing the round prevents interference of a read to other
read or write operations (of course a write can still interfere with a
read)