prbr: Fixed notification of original proposer on write through
Motivation of the mechanism is that the original proposer (A) of a
write is notified by the progress of the write through (B) if A has only
written a minority of replicas yet (And its requests was not yet fully
denied). This can prevent A from retrying its write when it receives
a majority of acks from B.

Previously A was stored as part of prbr:entry on each replica after A
has successfully written a value to it. When write through B tried to write on a
replica, A and B were notified. But in this setup A only receives acks from
replicas it has already written before (a minority).

Now, A is stored as part of the write through information of the write
round it used to try to write its value. Therefore every replica
receives this information when B starts a write through based on A's
partially finished write.

If A receives a write reply (ack or deny) from B's write through, all of A's
received replies (which have a smaller round) are obsolete and thrown
away. A then ignores all older replies it receives and waits for a
majority of acks/denies based on B's request.