prbr: repair mechanism for outdated replicas due to node failure
If one replica is empty (prbr_bottom) due to node failure and it
receives a write, it likely that the write is denied on this replica if
partial write operations are used (since all replicas must receive all
partial write ops in the same order).

If that is the case a repair mechanism can be triggered which will send a
qread request. The replica's data (rounds and value) will refreshed
using the return of the qread.