rbrcseq: Improved handling of reads executed concurrently to writes
This commit includes two protocol modifications to improve the handling
of reads that are executed concurrently to writes.

1) If the proposer of the read receives consistent write rounds it can
always deliver the read's result, even if the reads rounds are
inconsistent (which is a sign for an in-progress write). Due to the
consistent write rounds, the proposer can 'prove' that the write command has not
returned a result to a client yet, thus allowing it to safely deliver the old state
without violating strong consistency.

2) The use of command semantics to detect if the read commutes with
the write command. In such a case, the result of the read is independent
of the execution of the write, which allows the proposer to deliver the
read's result even if it does not know whether the write is already
done or not. Currently, rbrcseq checks the existence of an exported
function which can be used to define commutative readfilter,writefitler
pairs for the respective datatype.