rbrcseq: Reworked request state management of proposer
Proposer store the state of open requests locally. Arrives a message
from an acceptor, then the respecitve request-state is updated.
Previously, these entries were removed immediately once a quorum
decision was made (accept or deny). All arriving  message which belong
to a deleted request were ignored. However, it is useful to keep
specific request entries around longer in case later arriving messages
are of importance:

Concurrent write requests might lead to a denied write request which has
written a minority of replicas. Is this inconsistent state read by some
(other) proposer, then a write-through is started. Acceptors which vote
for this write-through write will notify the original proposer of their
vote. This way the original proposer can learn if its write was
successfull despide its initial deny. This is important to prevent the
same operation from being applied twice on the replicated state. This
change prevents that the original proposer removes its request state too
early for it to learn of the successfull write-through.

As long as the messages sent from a given acceptor to a proposer arrive
in the order they were sent in, the problem of applying the same write
twice is solved. However, this cannot be always guarantted (message loss,
transfered ownership of data to different node). Thus, this fix
only greatly reduces the probability of it from occuring.

A management entry was introduced which tracks the open request states
for a given client-submitted operation to reliably cleanup all relevant
requests once the operation was completed.