rbrcseq: Improved prevention of applying writes twice
If the original proper (OP) of a failed requests receives a success reply
from a write through, it knowns that some other proposer currently tries
to establish the new state. Since the OP receives replies from all
acceptors, it can behave as if itself has started the WT attempt
(i.e. continue the protocol form there if enough acceptors have replies
to make a decision). To prevent a double application of the write command,
eventual already started retries of OP's write are removed.

As long as only a minority of messages sent from acceptors to OP are lost or
delivered out-of-order, it is guaranteed that OP has not sent write
messages before the request is removed. Thus, it will not propose its
command a second time.