use db directory per vmname
to run without distributed Erlang, the db_directory has to become
independent of the erlang:node() name to avoid dile clashes of
different vms. Unfortunately, mnesia only works without distributed
erlang with the default node name 'nonode@nohost' correctly and does
not accept another fake node name as it then tries to do an rpc call
to this node, which fails. To circumvent this limitation, we create a
separate data sub-directory per vm. To do so, the db_directory given
in the config is extended by the value of the config entry
'vmname' (default 'novmname') on initialization. It can, for example
be given with '-e "-scalaris vmname my_vm"' to the scalarisctl script.

The particular db_* implementations still can use erlang:node() (aka
nonode@nohost) in their implementation.

TS&FS