prbr: do not use incremental rounds if some fixed round is known
This is part of a previous commit. However, I missed a sneaky occurence
when the phase 2 fails (because retry was handled by using retrigger).

The reason for not retrying with incremental
rounds is that many incremental mesages can be build-up if message
delivery is chaotic (e.g. proto shed without ordered links). This makes
it harder and harder for requests to finish as time progresses even
if a constant number of concurrent requests are handled.