prbr: Fixed bug that allows update to be applied multiple times
The previous mechanism was not sufficient to prevent this problem,
because it is possible for a proposer to retry a request even though its
proposal was chosen by a majority (without write-through):

-----
There are three acceptors. Acceptor 1 and 2 vote for proposal A (made by
proposer A). Acceptor 3 denies, because its round is too high.  However,
if A receives replies from acceptor 1 and 3 first, it must retry its
request. Otherwise it would risk to wait forever if 2 failed.

Before A retries, proposer B observes a consistent quorum by receiving
replies from 1 and 2. B proposes a follow-up value and suceeds. Now A
retries, but it has no information that its request suceeded. Thus, its
applies its update a second time even though no write-through took place!
----

To prevent this from happening, the previous mechanism can be modified
slightly. Basically, a proposer P sends in phase 2 the proposer P_old of the
previous chosen value. Acceptors now send two messages when voting. One vote message
to P as normal. And a "learned" message to P_old. Once P_old receives
the learned message it can know for certain that this request completed.
Reason for that is that a proposal must have been chosen before a
subsequent proposal can be proposed on top of it (consistent quorums). Exception to
this are write-throughs. In a write-through, acceptors do not send the
learned message.

Edgecase: A proposer P retries and receive a conistent quorum. It itself
made the previous proposal. Here, P can check if this was a proposal of
a previous attempt. If yes, then it sends a learned message to itself
before proposing something.

This mechanism guarantees that a proposer will receive a learned message
before it attempts to rety its proposal if a previous attempt already
succeeded. It does not matter if the original proposal was chosen normally or
because of a write-through. Upon receiption of a learned message, a
proposer stops retrying if its request is not done and notifies its client
Thus, no update is applied twice.