gla: reduced communication overhead between acceptor and proposer
proposer/acceptor only send commands that they have not previously
broadcasted, instead of the whole set. this assumes that messages
are not lost (as assumed by the paper), as no retry mechanism is implemented.
this adds storage overhead on the acceptor side as each acceptor has to
keep track which commands it already received from which proposer.

acceptors still send the full command set to the learners