prbr: Fixed premature write success client notification
Fixed bug in which an in-progress write received a learned message
caused by a successful write-through even though it is not completed.

This problem was caused due to the possibility of two writes using the
same round number (with different IDs) in the write phase. This happened
because proposer sometimes used a higher round number in the write phase
than their accepted round number in the read phase (i.e. if they received a
deny with a higher round number from a replica whilst collecting
replies of the first phase).

Observing two proposals with the same round number that where
each accepted by some replicas is problematic as it cannot always be
differentiated which one was established. This lead to the confusion
of sending the learned message to the wrong writer.